move logError to Logger.error
